django-auth-backend/requirements.txt
# WebQX Django Requirements
# Production-ready packages for global healthcare authentication

# Core Django
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9
django-redis==5.4.0

# Authentication & Security
django-allauth==0.57.0
django-otp==1.2.4
django-cors-headers==4.3.1
pyotp==2.9.0
qrcode[pil]==7.4.2

# Forms & UI
django-crispy-forms==2.1
crispy-bootstrap5==0.7

# Utilities
django-extensions==3.2.3
python-decouple==3.8
user-agents==2.2.0

# Background Tasks
celery==5.3.4
django-celery-beat==2.5.0
redis==5.0.1

# Email
django-ses==3.5.0

# Monitoring & Logging
django-health-check==3.17.0
sentry-sdk[django]==1.38.0

# Static Files
whitenoise==6.6.0

# Security
django-security==0.18.0
django-ratelimit==4.1.0

# Development Tools
django-debug-toolbar==4.2.0
django-silk==5.0.4

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Deployment
gunicorn==21.2.0
uvicorn[standard]==0.24.0

# AWS Integration
boto3==1.34.0
django-storages==1.14.2

# Utilities
python-dateutil==2.8.2
requests==2.31.0
Pillow==10.1.0

# API Documentation
drf-spectacular==0.26.5

# Production WSGI Server
gevent==23.9.1
